+++ web-server/src/main/java/com/se14/repository/ProjectRepository.java
@@ -7,7 +7,7 @@ import java.util.Optional;
 
 public interface ProjectRepository {
     Project save(Project project);
-    Optional<Project> findById(long id);
+    Optional<Project> findById(Integer id);
     List<Project> searchByTitle(String title); // returns all projects have title as substring
     List<Project> findAll();
 }
