+++ web-server/src/main/java/com/se14/repository/UserRepository.java
@@ -7,7 +7,7 @@ import java.util.Optional;
 
 public interface UserRepository {
     User save(User user);
-    Optional<User> findById(long id);
+    Optional<User> findById(Integer id);
     User findByUsername(String username);
     List<User> findAll();
 }
